Use visibility instead of opacity on stack
Setting and changing opacity to control the different
views of the stack led to the unintended behavior of
the cursor becoming a pointer when hovering over
pancakes that were hidden by opacity: 0 when numbers-
only mode was enabled. Visiblity removes this undesired
behavior and still maintains all functionality of the
stack.